refactored backend
