+++ backend/mongodb.js
@@ -1,4 +1,5 @@
 import { MongoClient } from 'mongodb';
+import { getMostCurrentQuarter } from './getMostCurrentQuarter';
 
 // Connection URL
 const url = process.env.MONGO_URL;
@@ -11,7 +12,7 @@ await usersCollection.createIndex({ uid: 1 }, { unique: true });
 
 // eslint-disable-next-line import/prefer-default-export
 
-db.collection('courses_20214').createIndex(
+db.collection(`courses_${await getMostCurrentQuarter()}`).createIndex(
   {
     courseID: 'text',
     title: 'text',
